# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Visca Corporation API",
        "description": "Dokumentasi Open Api Visca Corporation",
        "contact": {
            "email": "bimasaktiputra95@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server Local"
        }
    ],
    "paths": {
        "/v1/kategori/list": {
            "get": {
                "tags": [
                    "Kategori"
                ],
                "summary": "List Data Kategori Produk",
                "description": "List Data Kategori Produk",
                "operationId": "List Data Kategori Produk",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tipe pencarian berdasarkan nama kategori produk",
                        "required": false,
                        "schema": {
                            "type": "text"
                        },
                        "example": "Kebutuhan Pokok"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Data Kategori Produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": [
                                                    "List Data Kategori Produk"
                                                ]
                                            },
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/kategori/tambah-baru": {
            "post": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Tambah data kategori produk baru",
                "description": "Tambah data kategori produk baru",
                "operationId": "Tambah data kategori produk baru",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama Kategori Produk Baru",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil menambahkan kategori baru",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil menambahkan kategori baru"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Penambahan kategori produk gagal, periksa kembali data validasi anda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Penambahan kategori produk gagal, cek kembali validasi anda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/kategori/update-data/{id}": {
            "put": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Update data kategori produk",
                "description": "Update data kategori produk",
                "operationId": "Update data kategori produk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id kategori produk",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama kategori produk baru, yang akan menggantikan nama kategori lama",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil update data kategori produk",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil update data kategori produk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data kategori tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data kategori tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/kategori/hapus-data/{id}": {
            "delete": {
                "tags": [
                    "Kategori"
                ],
                "summary": "Hapus data kategori produk",
                "description": "Hapus data kategori produk",
                "operationId": "Hapus data kategori produk",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id data kategori produk yang akan di hapus",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Berhasil menghapus data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil menghapus data kategori"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Terjadi kesalahan saat menghapus data kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Terjadi kesalahan saat menghapus data kategori"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data produk tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolval",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data produk tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}